requirements.txt
google-generativeai
opencv-python
pillow
matplotlib
tqdm
python-dotenv
python-Levenshtein
ipykernel
jupyter
